feature: Binary Testing
This commit enables the feature toggle for Binary Testing - which switches to using the
real Terraform Binary when running Acceptance Tests and allows for different versions of
Core to be used.

Due to the use of Service Packages, we require that Reattach testing is used - as such
this requires using Terraform Core 0.12.26 or later.

This commit depends on - and must be rebased on top of - #10521 and #10519 due to the
removal of `acceptance.AzureProvider` when Binary Testing is being used.

A future commit will remove support for the legacy testing framework after a suitable
transition period, but for the moment disabling this functionality is sufficient.